+++ kleiner-brauhelfer-core/importexport.h
@@ -14,6 +14,7 @@ public:
     static int importMaischeMalzundMehr(Brauhelfer* bh, const QByteArray &content);
     static int importBeerXml(Brauhelfer* bh, const QByteArray &content);
     static QByteArray exportKbh(Brauhelfer* bh, int sudRow, const QStringList &exclude = QStringList());
+    static QByteArray exportBrautomat(Brauhelfer* bh, int sudRow);
     static QByteArray exportMaischeMalzundMehr(Brauhelfer* bh, int sudRow);
     static QByteArray exportBeerXml(Brauhelfer* bh, int sudRow);
 };